% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMuCoSoT_fit.R
\name{BayesMuCoSoT_fit}
\alias{BayesMuCoSoT_fit}
\title{Bayesian Multivariate Normal Modelling for Common Source Testing}
\usage{
BayesMuCoSoT_fit(
  y,
  x = NA,
  questioned_data,
  known_data,
  background_data = NA,
  background_data_id = NA,
  approach = "independent prior",
  BayesFactor_Approximation = "Bridge Sampling",
  prior_elicitation = "Maximum Likelihood",
  DoF = "min",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{y}{Vector of response columns names (p=variables).}

\item{x}{Predictors columns names (d=predictors). If there are any.}

\item{questioned_data}{Data frame with the unknown data.}

\item{known_data}{Data frame of the known data.}

\item{background_data}{Background data frame to estimate the prior parameters.
Background data recommended.}

\item{background_data_id}{Column name of the background data frame for
discriminate the data per id.}

\item{approach}{"conjugate" or "independent prior" approach.
Default "independent prior"}

\item{BayesFactor_Approximation}{"Laplace-Metropolis" or
"Generalized Harmonic Mean" or "Bridge Sampling". Default "Bridge Sampling".
 For the conjugate approach the Bayes factor comes in close form.}

\item{prior_elicitation}{"Non-Informative" or "Maximum Likelihood"
or list of type list(U= the inverse Wishart's scale matrix p by p,
beta_mu= matrix mean of the matrix Normal d by p,
beta_cov= d covariances matrix (3D matrix p by p by d).
Default "Maximum Likelihood" estimations from the Background data.}

\item{DoF}{Degrees of freedom for the prior inverse Wishart distribution.
Default "min": p+2. Otherwise, integer bigger than p+2.}

\item{verbose}{messages while running. Default TRUE.}

\item{...}{the other arguments of jags {R2jags} function.}
}
\value{
the logarithmic Bayes factor
}
\description{
Bayesian Multivariate Normal Modelling for Common Source Testing
}
