% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marg_Lik_functions.R
\name{marginal_likelihood_conjugate}
\alias{marginal_likelihood_conjugate}
\title{Conjugate Marginal Likelihood}
\usage{
marginal_likelihood_conjugate(jags_data)
}
\arguments{
\item{jags_data}{list of the names of the data objects used by the model (like jags data).}
}
\value{
the logarithmic marginal likelihood
}
\description{
Conjugate Marginal Likelihood
}
