% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marg_lik.R
\name{marginal_likelihood_niw_conjugate}
\alias{marginal_likelihood_niw_conjugate}
\title{Title}
\usage{
marginal_likelihood_niw_conjugate(w_data, m0, v0, k0, U0)
}
\arguments{
\item{U0}{}
}
\value{
Logarithmic Marginal Likelihood
}
\description{
Title
}
