% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marg_Lik_functions.R
\name{ml_bridge_sampling}
\alias{ml_bridge_sampling}
\title{Bridge Sampling  Marginal Likelihood Approximation}
\usage{
ml_bridge_sampling(samples, jags_data)
}
\arguments{
\item{samples}{samples of jags output (output$BUGSoutput$sims.matrix).}

\item{jags_data}{list of the names of the data objects used by the model (like jags data).}
}
\value{
the logarithmic marginal likelihood
}
\description{
Bridge Sampling  Marginal Likelihood Approximation
}
