% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marg_Lik_functions.R
\name{ml_generalized_harmonic_mean}
\alias{ml_generalized_harmonic_mean}
\title{Generalized Harmonic Mean  Marginal Likelihood Approximation}
\usage{
ml_generalized_harmonic_mean(samples, jags_data)
}
\arguments{
\item{samples}{samples of jags output (output$BUGSoutput$sims.matrix).}

\item{jags_data}{list of the names of the data objects used by the model (like jags data).}
}
\value{
the logarithmic marginal likelihood
}
\description{
Generalized Harmonic Mean  Marginal Likelihood Approximation
}
